Suporte a notas

+++ lib.php
@@ -37,7 +37,7 @@ function supervideo_supports($feature) {
         case FEATURE_SHOW_DESCRIPTION:
             return true;
         case FEATURE_GRADE_HAS_GRADE:
-            return false;
+            return true;
         case FEATURE_GRADE_OUTCOMES:
             return true;
         case FEATURE_BACKUP_MOODLE2:
@@ -46,11 +46,90 @@ function supervideo_supports($feature) {
             return true;
         case FEATURE_COMPLETION_HAS_RULES:
             return true;
-        case 'mod_purpose':
-            return 'content';
         default:
             return null;
     }
+}
+
+/**
+ * @param      $supervideo
+ * @param null $grades
+ *
+ * @return int
+ */
+function supervideo_grade_item_update($supervideo, $grades = null) {
+    global $CFG;
+
+    require_once($CFG->libdir . '/gradelib.php');
+
+    $params = [
+        'itemname' => $supervideo->name,
+        'idnumber' => $supervideo->cmidnumber,
+        'gradetype' => GRADE_TYPE_VALUE,
+        'grademax' => 100,
+        'grademin' => 0
+    ];
+
+    if ($grades === 'reset') {
+        $params['reset'] = true;
+        $grades = null;
+    }
+
+    return grade_update('mod/supervideo', $supervideo->course, 'mod', 'supervideo', $supervideo->id, 0, $grades, $params);
+}
+
+/**
+ * @param      $supervideo
+ * @param int  $userid
+ * @param bool $nullifnone
+ *
+ * @return null
+ * @throws coding_exception
+ * @throws dml_exception
+ */
+function supervideo_update_grades($supervideo, $userid = 0, $nullifnone = true) {
+    global $CFG;
+    require_once($CFG->libdir . '/gradelib.php');
+
+    if (!$supervideo->grade_approval) {
+        return null;
+    }
+
+    if ($grades = supervideo_get_user_grades($supervideo, $userid)) {
+        supervideo_grade_item_update($supervideo, $grades);
+    }
+}
+
+/**
+ * @param     $supervideo
+ * @param int $userid
+ *
+ * @return array|bool
+ * @throws coding_exception
+ * @throws dml_exception
+ */
+function supervideo_get_user_grades($supervideo, $userid = 0) {
+    global $DB;
+
+    if (!$supervideo->grade_approval) {
+        return false;
+    }
+
+    $cm = get_coursemodule_from_instance('supervideo', $supervideo->id);
+
+    $params = ['cm_id' => $cm->id];
+
+    $extrawhere = ' ';
+    if ($userid > 0) {
+        $extrawhere .= ' AND user_id = :user_id';
+        $params['user_id'] = $userid;
+    }
+
+    $sql = "SELECT user_id as userid, MAX(percent) as rawgrade
+              FROM {supervideo_view}
+             WHERE cm_id = :cm_id {$extrawhere}
+             GROUP BY user_id";
+    return $DB->get_records_sql($sql, $params);
 }
 
 /**
@@ -62,11 +141,13 @@ function supervideo_supports($feature) {
  * @throws coding_exception
  */
 function supervideo_add_instance(stdClass $supervideo, mod_supervideo_mod_form $mform = null) {
-    global $DB, $USER, $OUTPUT;
+    global $DB;
 
+    $supervideo->timemodified = time();
     $supervideo->timecreated = time();
     $supervideo->id = $DB->insert_record('supervideo', $supervideo);
 
+    supervideo_grade_item_update($supervideo);
     supervideo_set_mainfile($supervideo);
 
     return $supervideo->id;
@@ -84,14 +165,14 @@ function supervideo_set_mainfile($supervideo) {
 
     $context = context_module::instance($cmid);
     if ($draftitemid) {
-        file_save_draft_area_files($draftitemid, $context->id, 'mod_supervideo', 'content', 0);
+        $options = ['subdirs' => true, 'embed' => true];
+        file_save_draft_area_files($draftitemid, $context->id, 'mod_supervideo', 'content', $supervideo->id, $options);
     }
-//    $files = $fs->get_area_files($context->id, 'mod_supervideo', 'content', 0, 'sortorder', false);
-//    if (count($files) == 1) {
-//        $file = reset($files);
-//        file_set_sortorder($context->id, 'mod_supervideo', 'content', 0, $file->get_filepath(), $file->get_filename(), 1);
-//    }
-
+    $files = $fs->get_area_files($context->id, 'mod_supervideo', 'content', 0, 'sortorder', false);
+    if (count($files) == 1) {
+        $file = reset($files);
+        file_set_sortorder($context->id, 'mod_supervideo', 'content', 0, $file->get_filepath(), $file->get_filename(), 1);
+    }
 }
 
 /**
@@ -111,6 +192,8 @@ function supervideo_update_instance(stdClass $supervideo, mod_supervideo_mod_for
 
     $result = $DB->update_record('supervideo', $supervideo);
 
+    supervideo_grade_item_update($supervideo);
+
     return $result;
 }
 
@@ -129,14 +212,16 @@ function supervideo_delete_instance($id) {
     if (!$supervideo = $DB->get_record('supervideo', array('id' => $id))) {
         return false;
     }
+
     $fs = get_file_storage();
     $cm = get_coursemodule_from_id('supervideo', $supervideo->id);
-    $files = $fs->get_area_files(context_module::instance($cm->id)->id, 'mod_supervideo', 'content', $supervideo->id, 'sortorder DESC, id ASC', false);
+    if ($cm) {
+        $files = $fs->get_area_files(context_module::instance($cm->id)->id, 'mod_supervideo', 'content', $supervideo->id, 'sortorder DESC, id ASC', false);
 
-    foreach ($files as $file) {
-        $file->delete();
+        foreach ($files as $file) {
+            $file->delete();
+        }
     }
-
     $DB->delete_records('supervideo', array('id' => $supervideo->id));
     $DB->delete_records('supervideo_view', array('cm_id' => $cm->id));
 
@@ -288,6 +373,11 @@ function supervideo_extend_navigation_course($navigation, $course, $context) {
     }
 }
 
+function supervideo_get_completion_state($course, $cm, $userid, $type) {
+    return \mod_supervideo\grades::supervideo_get_completion_state($cm);
+}
+
+
 /**
  * Serve the files from the supervideo file areas
  *
@@ -347,4 +437,115 @@ function supervideo_pluginfile($course, $cm, context $context, $filearea, $args,
 
     // We can now send the file back to the browser - in this case with a cache lifetime of 1 day and no filtering.
     send_stored_file($file, 86400, 0, $forcedownload, $options);
+}
+
+/**
+ * Register the ability to handle drag and drop file uploads
+ * @return array containing details of the files / types the mod can handle
+ * @throws coding_exception
+ */
+function _supervideo_dndupload_register() {
+    $ret = [
+        'files' => [
+            'extension' => '.mp4',
+            'message' => get_string('dnduploadlabel-mp4', 'mod_supervideo')
+        ],
+        [
+            'extension' => '.mp3',
+            'message' => get_string('dnduploadlabel-mp3', 'mod_supervideo')
+        ]
+    ];
+
+    $strdndtext = get_string('dnduploadlabeltext', 'mod_supervideo');
+    return array_merge($ret, array('types' => array(
+        ['identifier' => 'text/html', 'message' => $strdndtext, 'noname' => true],
+        ['identifier' => 'text', 'message' => $strdndtext, 'noname' => true]
+    )));
+}
+
+/**
+ * Register the ability to handle drag and drop file uploads
+ *
+ * @return array containing details of the files / types the mod can handle
+ * @throws coding_exception
+ */
+function supervideo_dndupload_register() {
+    $ret = [
+        'files' => [
+            [
+                'extension' => 'mp3',
+                'message' => get_string('dnduploadlabel-mp3', 'mod_supervideo'),
+            ],
+            [
+                'extension' => 'mp4',
+                'message' => get_string('dnduploadlabel-mp4', 'mod_supervideo'),
+            ]
+        ],
+        'types' => [
+            [
+                'identifier' => 'text/html',
+                'message' => get_string('dnduploadlabeltext', 'mod_supervideo'),
+                'noname' => true
+            ],
+            [
+                'identifier' => 'text',
+                'message' => get_string('dnduploadlabeltext', 'mod_supervideo'),
+                'noname' => true
+            ]
+        ]
+    ];
+    return $ret;
+}
+
+
+/**
+ * Handle a file that has been uploaded
+ *
+ * @param object $uploadinfo details of the file / content that has been uploaded
+ *
+ * @return int instance id of the newly created mod
+ * @throws coding_exception
+ * @throws dml_exception
+ */
+function supervideo_dndupload_handle($uploadinfo) {
+    global $USER;
+
+    // Gather the required info.
+    $data = new stdClass();
+    $data->course = $uploadinfo->course->id;
+    $data->name = $uploadinfo->displayname;
+    $data->intro = '';
+    $data->introformat = FORMAT_HTML;
+    $data->coursemodule = $uploadinfo->coursemodule;
+
+    $data->videosize = 1;
+    $data->showrel = 0;
+    $data->showcontrols = 1;
+    $data->showshowinfo = 0;
+    $data->autoplay = 0;
+    $data->grade_approval = 0;
+
+    $data->instance = supervideo_add_instance($data, null);
+
+    if (!empty($uploadinfo->draftitemid)) {
+        $fs = get_file_storage();
+        $draftcontext = context_user::instance($USER->id);
+        $context = context_module::instance($uploadinfo->coursemodule);
+        $files = $fs->get_area_files($draftcontext->id, 'user', 'draft', $uploadinfo->draftitemid, '', false);
+        if ($file = reset($files)) {
+
+            $data->videourl = "[resource-file:{$file->get_filename()}]";
+            $options = ['subdirs' => true, 'embed' => true];
+            file_save_draft_area_files($uploadinfo->draftitemid, $context->id, 'mod_supervideo', 'content', $data->instance, $options);
+
+            supervideo_update_instance($data, null);
+        }
+    } else if (!empty($uploadinfo->content)) {
+        $data->intro = $uploadinfo->content;
+        if ($uploadinfo->type != 'text/html') {
+            $data->introformat = FORMAT_PLAIN;
+        }
+    }
+
+    return $data->instance;
 }
